requirements.txt
fastapi
uvicorn
pydantic
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2